Create overflowLine component to use in NormalRightRow
Refactored the displayDescriptor method within NormalRightRow into its
own component OverflowLine.
